key_vault/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
asgiref==3.10.0
sqlparse==0.5.3
